+++ src/index.ts
@@ -218,5 +218,4 @@ export { PermissionUtil } from "./models/v2/resources/permission-util";
 export { IPermissions } from "./interfaces/models/admin/i-permissions";
 export { IFulltextSearchParams } from "./interfaces/models/v2/i-fulltext-search-params";
 export { ILabelSearchParams } from "./interfaces/models/v2/i-label-search-params";
-export { IUrl } from "./models/v2/custom-converters/base-url-converter";
-export { IId } from "./models/v2/custom-converters/union-data-management-plan-id-converter";
+export { IId, IUrl } from "./models/v2/project-metadata/metadata-interfaces";
