+++ src/models/v2/project-metadata/metadata-interfaces.ts
@@ -1,0 +1,10 @@
+export interface IId {
+    id: string;
+}
+
+export interface IUrl {
+    type?: string;
+    value?: string;
+    name?: string;
+    url?: string;
+}
