+++ src/models/v2/project-metadata/organization.ts
@@ -1,8 +1,8 @@
 import { JsonObject, JsonProperty } from "json2typescript";
 import { Constants } from "../Constants";
-import { IUrl } from "../custom-converters/base-url-converter";
 import { UnionStringArrayOfStringsConverter } from "../custom-converters/union-string-array-of-strings-converter";
 import { UrlToUrlObjectConverter } from "../custom-converters/url-to-url-object-converter";
+import { IUrl } from "../project-metadata/metadata-interfaces";
 import { Address } from "./address";
 import { BaseProjectMetadata } from "./base-project-metadata";
