+++ src/models/v2/project-metadata/person.ts
@@ -1,10 +1,9 @@
 import { JsonObject, JsonProperty } from "json2typescript";
 import { Constants } from "../Constants";
-import { IUrl } from "../custom-converters/base-url-converter";
-import { IId } from "../custom-converters/union-data-management-plan-id-converter";
 import { UnionOrganizationIdConverter } from "../custom-converters/union-orgnization-id-converter";
 import { UnionStringArrayOfStringsConverter } from "../custom-converters/union-string-array-of-strings-converter";
 import { UrlToUrlObjectConverter } from "../custom-converters/url-to-url-object-converter";
+import { IId, IUrl } from "../project-metadata/metadata-interfaces";
 import { Address } from "./address";
 import { BaseProjectMetadata } from "./base-project-metadata";
 import { Organization } from "./organization";
