+++ test-framework/src/app/app.component.ts
@@ -58,11 +58,9 @@ import {
     ProjectsMetadata,
     Dataset,
     SingleProject,
-    DataManagementPlan,
     UpdateProjectMetadataResponse,
     IUrl,
     Grant,
-    Person,
     Organization,
     UpdateChildNodeRequest,
     ListNodeInfoResponse,
@@ -73,7 +71,6 @@ import {
     UpdateResourcePropertyLabel,
     UpdateResourcePropertyComment
 } from '@dasch-swiss/dsp-js';
-import { IId } from '@dasch-swiss/dsp-js/src/models/v2/custom-converters/union-data-management-plan-id-converter';
 import { Observable } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
